Add refs and extra scope to intersect1d which has the following impacts:

Memory: significant improvements to the high mark of memory
`Memory.memoryUsed()` with 10,000 elements on a single locale 
| time of measure | master | improvement |
| ----------- | -----: | ----: |
| mem at beginning     | 310280 | 310280 |
| mem high mark   | 924285 | 618815 |

Performance: slight speed increase
Execution time of 1,00,000,000 elements on 16 locales in seconds (lower is better)
|  | master | improvement |
| ----------- | -----: | ----: |
| execution time (s) | 2.53 | 2.31 |